Saving partial corpus progress
New issue with the FM corpus…the scrape was successful but it took a
couple of hours…which is to be expected. That said, there was a time
out issue with RStudio that cause the connection to R to lost. I am now
running the same retrieval loop but through the primative R interface
to get around the timeout/disconnecting issue (which seems inherent to
any long running loops or programs in RStudio)